Added code for loading spinner and moved all modal forms into a directory
