+++ client/components/modal-components/studyoptions.jsx
