+++ client/pages/editcard.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Modal from '../components/modal';
 import AppContext from '../lib/app-context';
-import DeleteForm from '../components/deletecard';
+import DeleteForm from '../components/modal-components/deletecard';
 
 export default class EditCard extends React.Component {
   constructor(props) {
